+++ aiogram_dialog/__init__.py
@@ -1,6 +1,6 @@
-from . import exceptions
-from .context.events import ChatEvent, Data, ShowMode, StartMode
-from .context.stack import DEFAULT_STACK_ID
+from .api import exceptions
+from .api.events import ChatEvent, Data, ShowMode, StartMode
+from .api.stack import DEFAULT_STACK_ID
 from .dialog import Dialog
 from .manager.protocols import BaseDialogManager, DialogManager
 from .manager.registry import DialogRegistry
