+++ aiogram_dialog/api/exceptions.py
