+++ aiogram_dialog/context/intent_filter.py
@@ -7,11 +7,13 @@ from aiogram.fsm.state import StatesGroup
 from aiogram.fsm.storage.base import BaseStorage
 from aiogram.types import CallbackQuery, Message, TelegramObject, Update
 
-from .context import Context
-from .events import ChatEvent, DialogUpdate, DialogUpdateEvent
-from .stack import Stack
+from aiogram_dialog.api.context import Context
+from aiogram_dialog.api.events import (
+    ChatEvent, DialogUpdate, DialogUpdateEvent,
+)
+from aiogram_dialog.api.exceptions import InvalidStackIdError, OutdatedIntent
+from aiogram_dialog.api.stack import Stack
 from .storage import DEFAULT_STACK_ID, StorageProxy
-from ..exceptions import InvalidStackIdError, OutdatedIntent
 from ..utils import remove_indent_id
 
 STORAGE_KEY = "aiogd_storage_proxy"
