+++ aiogram_dialog/dialog.py
@@ -17,8 +17,8 @@ from aiogram.exceptions import TelegramAPIError
 from aiogram.fsm.state import State, StatesGroup
 from aiogram.types import CallbackQuery, InlineKeyboardMarkup, Message
 
-from .context.events import Data
-from .exceptions import UnregisteredWindowError
+from aiogram_dialog.api.events import Data
+from aiogram_dialog.api.exceptions import UnregisteredWindowError
 from .manager.protocols import (
     DialogManager,
     DialogRegistryProto,
